+++ common/routes/Explore/components/ExploreError.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const ExploreError = props => (
+  <div>
+    Explore Error
+  </div>
+);
+
+export default ExploreError;
