+++ common/routes/Explore/components/ExploreFeed.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const ExploreFeed = props => (
+  <div>
+    Explore Feed
+  </div>
+);
+
+export default ExploreFeed;
